Parser: removed segfault in syntax_jumptable
