+++ src/parser/syntax_jmptbl_func.c
@@ -1,0 +1,36 @@
+#include <minishell.h>
+
+bool	syntax_id_quote(const t_token *n_prev, const t_token *n_cur)
+{
+	(void) n_prev;
+	(void) n_cur;
+	return (0);
+}
+
+bool	syntax_id_pipe(const t_token *n_prev, const t_token *n_cur)
+{
+	(void) n_prev;
+	(void) n_cur;
+	return (0);
+}
+
+bool	syntax_id_redir(const t_token *n_prev, const t_token *n_cur)
+{
+	(void) n_prev;
+	(void) n_cur;
+	return (0);
+}
+
+bool	syntax_id_space(const t_token *n_prev, const t_token *n_cur)
+{
+	(void) n_prev;
+	(void) n_cur;
+	return (0);
+}
+
+bool	syntax_id_word(const t_token *n_prev, const t_token *n_cur)
+{
+	(void) n_prev;
+	(void) n_cur;
+	return (0);
+}
